# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "新能源材料进销存 API",
    "description": "进销存后端接口：用户与库房、收货定价、对外报价、入库审核、出库与 FIFO、库存报表与预警等。默认账号见服务启动日志。详细说明见仓库内 `server/API接口文档.md`。\n\n**鉴权**：除「健康检查」「登录」「自助注册」外，请求头需 `Authorization: Bearer <JWT>`。角色 `statistics`（统计部）与 `warehouse`（库房）权限不同，见各接口说明。\n\n**Swagger**：展开接口后可用 **Try it out** 调试；**Curl** 区会按当前参数生成命令；**Example Value** 为示例请求/响应体。",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "当前服务（与浏览器同源）" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 token"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录或 Token 无效",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" },
            "example": { "error": "未登录或缺少 Token" }
          }
        }
      },
      "Forbidden": {
        "description": "无权限",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" },
            "example": { "error": "仅统计部可操作" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/weighbridge-slip/parse": {
      "post": {
        "tags": ["出库与磅单"],
        "summary": "解析磅单 OCR 文本（库房）",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "imageUrl": "https://example.com/pound.jpg",
                "ocrText": "品名 废钢 净重 25.66 吨"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "解析结果（与 weighbridgeParse 一致）",
            "content": {
              "application/json": {
                "example": {
                  "imageUrl": "https://example.com/pound.jpg",
                  "recognized": true,
                  "suggestedMaterialId": 1,
                  "suggestedWeight": 25.66,
                  "numbersFound": [25.66, 30],
                  "candidates": [
                    { "materialId": 1, "code": "MAT-001", "name": "示例品种", "matchBy": "name" }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/admin/outbound-orders": {
      "get": {
        "tags": ["出库与磅单"],
        "summary": "出库单分页列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } },
          { "name": "materialId", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "completed"] } }
        ],
        "responses": {
          "200": {
            "description": "列表",
            "content": {
              "application/json": {
                "example": {
                  "orders": [
                    {
                      "id": 1,
                      "orderNo": "CK-20260424-0001",
                      "warehouseId": 1,
                      "materialId": 1,
                      "plannedWeight": 5,
                      "unitPrice": 3500,
                      "status": "pending"
                    }
                  ],
                  "total": 1,
                  "page": 1,
                  "pageSize": 10
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["出库与磅单"],
        "summary": "创建出库单（库房，FIFO 子行自动分配）",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "warehouseId": 1,
                "materialId": 1,
                "plannedWeight": 5,
                "unitPrice": 3500
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "含 fifoLines",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "orderNo": "CK-20260424-0001",
                  "plannedWeight": 5,
                  "status": "pending",
                  "fifoLines": [
                    {
                      "inboundOrderId": 1,
                      "plannedWeight": 5,
                      "lineNo": 1,
                      "subOrderNo": "CK-20260424-0001-S001"
                    }
                  ],
                  "defaultOutboundUnitPrice": 3500
                }
              }
            }
          },
          "400": {
            "description": "库存不足等",
            "content": {
              "application/json": {
                "example": {
                  "error": "可出库库存不足（先进先出），请减少预出库重量或先完成入库审核",
                  "shortfall": 2.5
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/admin/outbound-orders/{id}": {
      "get": {
        "tags": ["出库与磅单"],
        "summary": "出库单详情",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "含 fifoLines、defaultOutboundUnitPrice",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "orderNo": "CK-20260424-0001",
                  "fifoLines": [],
                  "defaultOutboundUnitPrice": 3500
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "不存在",
            "content": {
              "application/json": { "example": { "error": "出库单不存在" } }
            }
          }
        }
      },
      "delete": {
        "tags": ["出库与磅单"],
        "summary": "删除待完成出库单（库房）",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "已删除" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/admin/outbound-orders/{id}/complete": {
      "put": {
        "tags": ["出库与磅单"],
        "summary": "确认实际出库（库房）",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "actualWeight": 4.98,
                "weighbridgePhoto": "https://example.com/bridge.jpg"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "完成后详情",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "status": "completed",
                  "actualWeight": 4.98,
                  "weighbridgePhoto": "https://example.com/bridge.jpg",
                  "fifoLines": []
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  }
}